requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_tflow_select=2.2.0=eigen
absl-py=0.10.0=py37_0
argon2-cffi=20.1.0=py37he774522_1
astor=0.8.1=py37_0
attrs=20.2.0=py_0
backcall=0.2.0=py_0
blas=1.0=mkl
bleach=3.2.1=py_0
blinker=1.4=py37_0
brotlipy=0.7.0=py37he774522_1000
ca-certificates=2020.7.22=0
cachetools=4.1.1=py_0
certifi=2020.6.20=py37_0
cffi=1.14.3=py37h7a1dbc1_0
chardet=3.0.4=py37_1003
click=7.1.2=py_0
cloudpickle=1.6.0=py_0
colorama=0.4.3=py_0
cryptography=3.1=py37h7a1dbc1_0
cudatoolkit=10.2.89=h74a9793_1
cycler=0.10.0=py37_0
cytoolz=0.10.1=py37he774522_0
dask-core=2.27.0=py_0
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
entrypoints=0.3=py37_0
freetype=2.10.2=hd328e21_0
gast=0.2.2=py37_0
google-auth=1.21.2=py_0
google-auth-oauthlib=0.4.1=py_2
google-pasta=0.2.0=py_0
grpcio=1.31.0=py37he7da953_0
h5py=2.8.0=py37hf7173ca_2
hdf5=1.8.20=hac2f561_1
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha925a31_3
idna=2.10=py_0
imageio=2.9.0=py_0
importlib-metadata=1.7.0=py37_0
importlib_metadata=1.7.0=0
intel-openmp=2020.2=254
ipykernel=5.3.2=py37h5ca1d4c_0
ipython=7.18.1=py37h5ca1d4c_0
ipython_genutils=0.2.0=py37_0
jedi=0.17.2=py37_0
jinja2=2.11.2=py_0
jpeg=9b=hb83a4c4_2
jsonschema=3.2.0=py37_1
jupyter_client=6.1.6=py_0
jupyter_core=4.6.3=py37_0
keras-applications=1.0.8=py_1
keras-preprocessing=1.1.0=py_1
kiwisolver=1.2.0=py37h74a9793_0
libopencv=3.4.2=h20b85fd_0
libpng=1.6.37=h2a8f88b_0
libprotobuf=3.13.0=h200bbdf_0
libsodium=1.0.18=h62dcd97_0
libtiff=4.1.0=h56a325e_1
lz4-c=1.9.2=h62dcd97_1
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markdown=3.2.2=py37_0
markupsafe=1.1.1=py37hfa6e2cd_1
matplotlib=3.2.2=0
matplotlib-base=3.2.2=py37h64f37c6_0
mistune=0.8.4=py37hfa6e2cd_1001
mkl=2020.2=256
mkl-service=2.3.0=py37hb782905_0
mkl_fft=1.2.0=py37h45dec08_0
mkl_random=1.1.1=py37h47e9c7a_0
msys2-conda-epoch=20160418=1
nb_conda_kernels=2.2.3=py37_0
nbconvert=5.6.1=py37_1
nbformat=5.0.7=py_0
networkx=2.5=py_0
ninja=1.10.1=py37h7ef1ec2_0
notebook=6.1.1=py37_0
numpy=1.19.1=py37h5510c5b_0
numpy-base=1.19.1=py37ha3acd2a_0
oauthlib=3.1.0=py_0
olefile=0.46=py37_0
opencv=3.4.2=py37h40b0b35_0
openssl=1.1.1g=he774522_1
opt_einsum=3.1.0=py_0
packaging=20.4=py_0
pandas=1.0.5=py37h47e9c7a_0
pandoc=2.10.1=0
pandocfilters=1.4.2=py37_1
parso=0.7.0=py_0
pickleshare=0.7.5=py37_1001
pillow=7.2.0=py37hcc1f983_0
pims=0.5=pyh9f0ad1d_1
pip=20.2.2=py37_0
prometheus_client=0.8.0=py_0
prompt-toolkit=3.0.7=py_0
protobuf=3.13.0=py37h6538335_0
py-opencv=3.4.2=py37hc319ecb_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.8=py_0
pycparser=2.20=py_2
pygments=2.7.1=py_0
pyjwt=1.7.1=py37_0
pyopenssl=19.1.0=py_1
pyparsing=2.4.7=py_0
pyqt=5.9.2=py37h6538335_2
pyrsistent=0.17.3=py37he774522_0
pysocks=1.7.1=py37_1
python=3.7.9=h60c2a47_0
python-dateutil=2.8.1=py_0
pytorch=1.6.0=py3.7_cuda102_cudnn7_0
pytz=2020.1=py_0
pywavelets=1.1.1=py37he774522_0
pywin32=227=py37he774522_1
pywinpty=0.5.7=py37_0
pyyaml=5.3.1=py37he774522_1
pyzmq=19.0.2=py37ha925a31_1
qt=5.9.7=vc14h73c81de_0
requests=2.24.0=py_0
requests-oauthlib=1.3.0=py_0
rsa=4.6=py_0
scikit-image=0.16.2=py37h47e9c7a_0
scipy=1.5.2=py37h9439919_0
send2trash=1.5.0=py37_0
setuptools=49.6.0=py37_0
sip=4.19.8=py37h6538335_0
six=1.15.0=py_0
slicerator=1.0.0=py_0
sqlite=3.33.0=h2a8f88b_0
tensorboard=2.2.1=pyh532a8cf_0
tensorboard-plugin-wit=1.6.0=py_0
tensorflow=2.1.0=eigen_py37hd727fc0_0
tensorflow-base=2.1.0=eigen_py37h49b2757_0
tensorflow-estimator=2.1.0=pyhd54b08b_0
termcolor=1.1.0=py37_1
terminado=0.8.3=py37_0
testpath=0.4.4=py_0
tifffile=0.15.1=py37h2a96729_2
tk=8.6.10=he774522_0
toolz=0.10.0=py_0
torchvision=0.7.0=py37_cu102
tornado=6.0.4=py37he774522_1
tqdm=4.48.2=py_0
traitlets=5.0.4=py_0
urllib3=1.25.10=py_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_3
wcwidth=0.2.5=py_0
webencodings=0.5.1=py37_1
werkzeug=0.16.1=py_0
wheel=0.35.1=py_0
win_inet_pton=1.1.0=py37_0
wincertstore=0.2=py37_0
winpty=0.4.3=4
wrapt=1.12.1=py37he774522_1
xz=5.2.5=h62dcd97_0
yaml=0.2.5=he774522_0
zeromq=4.3.2=ha925a31_3
zipp=3.1.0=py_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.5=h04227a9_0
